requirements.txt
Flask==1.0.3
Flask-Cors==3.0.7
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
Flask-Migrate==2.5.3
SQLAlchemy==1.3.4
psycopg2-binary==2.8.2
pycryptodome==3.6.6
pylint==2.3.1
python-jose-cryptodome==1.3.2
babel
python-dateutil==2.6.0
flask-moment
flask-wtf
flask_script
python-dotenv
gunicorn==20.0.4
Werkzeug==0.15.3 
wrapt==1.11.1
bleach==3.1.5